+++ packages/expo-updates/src/Updates.types.ts
@@ -1,4 +1,6 @@
-import Constants from 'expo-constants';
+import { NewManifest, BareManifest } from 'expo-manifests';
+
+export type Manifest = NewManifest | BareManifest;
 
 /**
  * The types of update-related events.
@@ -19,21 +21,6 @@ export enum UpdateEventType {
    */
   ERROR = 'error',
 }
-
-// @docsMissing
-// TODO(eric): move source of truth for manifest type to this module
-/**
- * @hidden
- */
-export type ClassicManifest = typeof Constants.manifest;
-
-// @docsMissing
-/**
- * @hidden
- */
-export type Manifest = ClassicManifest | typeof Constants.manifest2;
-// modern manifest type is intentionally not exported, since the plan is to call it just "Manifest"
-// in the future
 
 type UpdateCheckResultRollBackToEmbedded = {
   /**
@@ -272,6 +259,14 @@ export type LocalAssets = Record<string, string>;
 /**
  * @hidden
  */
+export type UpdatesNativeStateRollback = {
+  // ISO date string with the rollback commit time
+  commitTime: string;
+};
+
+/**
+ * @hidden
+ */
 export type UpdatesNativeStateMachineContext = {
   // The native state machine context, either read directly from a native module method,
   // or received in a state change event. Used internally by this module and not exported publicly.
@@ -279,10 +274,10 @@ export type UpdatesNativeStateMachineContext = {
   isUpdatePending: boolean;
   isChecking: boolean;
   isDownloading: boolean;
-  isRollback: boolean;
   isRestarting: boolean;
   latestManifest?: Manifest;
   downloadedManifest?: Manifest;
+  rollback?: UpdatesNativeStateRollback;
   checkError?: Error;
   downloadError?: Error;
   lastCheckForUpdateTime?: Date;