+++ packages/@expo/cli/src/start/server/webpack/__tests__/WebpackBundlerDevServer-test.ts
@@ -64,11 +64,11 @@ describe('startAsync', () => {
       messageSocket: {
         broadcast: expect.any(Function),
       },
-      middleware: undefined,
+      middleware: null,
       server: {
         close: expect.any(Function),
         listen: expect.any(Function),
-        sockWrite: expect.any(Function),
+        sendMessage: expect.any(Function),
       },
     });
 
@@ -106,19 +106,7 @@ describe('broadcastMessage', () => {
 
     expect(
       // @ts-expect-error
-      devServer.getInstance().server.sockWrite
+      devServer.getInstance().server.sendMessage
     ).toBeCalledWith(undefined, 'content-changed', { foo: true });
-  });
-  it(`uses custom handler`, async () => {
-    const devServer = await getStartedDevServer();
-    devServer['customMessageSocketBroadcaster'] = jest.fn();
-    devServer.broadcastMessage('reload', { foo: true });
-
-    expect(
-      // @ts-expect-error
-      devServer.getInstance().server.sockWrite
-    ).not.toBeCalled();
-
-    expect(devServer['customMessageSocketBroadcaster']).toBeCalledWith('reload', { foo: true });
   });
 });