+++ packages/expo-auth-session/src/__tests__/SessionUrlProvider-test.ts
@@ -2,7 +2,7 @@ import { ExecutionEnvironment } from 'expo-constants';
 import { Platform } from 'expo-modules-core';
 import { unmockAllProperties } from 'jest-expo';
 
-import { describeManifestTypes, mockConstants } from './ManifestTestUtils';
+import { describeManifest, mockConstants } from './ManifestTestUtils';
 
 beforeEach(() => {
   unmockAllProperties();
@@ -11,19 +11,16 @@ beforeEach(() => {
 });
 
 describe('getStartUrl', () => {
-  describeManifestTypes(
-    { id: '@test/test', originalFullName: '@example/abc', scheme: 'my-app' },
-    {
-      extra: {
-        expoClient: {
-          name: 'wat',
-          slug: 'wat',
-          originalFullName: '@example/abc',
-          scheme: 'my-app',
-        },
+  describeManifest({
+    extra: {
+      expoClient: {
+        name: 'wat',
+        slug: 'wat',
+        originalFullName: '@example/abc',
+        scheme: 'my-app',
       },
-    }
-  )((manifestObj) => {
+    },
+  })((manifestObj) => {
     it(`returns the correct start URL from getStartUrl`, () => {
       mockConstants({}, manifestObj);
 
@@ -86,19 +83,16 @@ describe(`getRedirectUrl`, () => {
       });
       afterEach(() => (console.warn = originalWarn));
 
-      describeManifestTypes(
-        { id: '@test/test', originalFullName: '@example/abc', scheme: 'my-app' },
-        {
-          extra: {
-            expoClient: {
-              name: 'wat',
-              slug: 'wat',
-              originalFullName: '@example/abc',
-              scheme: 'my-app',
-            },
+      describeManifest({
+        extra: {
+          expoClient: {
+            name: 'wat',
+            slug: 'wat',
+            originalFullName: '@example/abc',
+            scheme: 'my-app',
           },
-        }
-      )((manifestObj) => {
+        },
+      })((manifestObj) => {
         it(`checks return url`, () => {
           mockConstants({ executionEnvironment: execution }, manifestObj);
 
@@ -115,19 +109,16 @@ describe(`getRedirectUrl`, () => {
       });
 
       if (Platform.OS !== 'web') {
-        describeManifestTypes(
-          { id: undefined, originalFullName: undefined, scheme: 'my-app' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                originalFullName: undefined,
-                scheme: 'my-app',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              originalFullName: undefined,
+              scheme: 'my-app',
             },
-          }
-        );
+          },
+        });
       }
     });
   }
@@ -136,19 +127,16 @@ describe(`getRedirectUrl`, () => {
 if (Platform.OS !== 'web') {
   describe(`getDefaultReturnUrl`, () => {
     describe('storeClient', () => {
-      describeManifestTypes(
-        { scheme: 'my-app', hostUri: 'exp.host/@example/abc' },
-        {
-          extra: {
-            expoClient: {
-              name: 'wat',
-              slug: 'wat',
-              scheme: 'my-app',
-              hostUri: 'exp.host/@example/abc',
-            },
+      describeManifest({
+        extra: {
+          expoClient: {
+            name: 'wat',
+            slug: 'wat',
+            scheme: 'my-app',
+            hostUri: 'exp.host/@example/abc',
           },
-        }
-      )((manifestObj) => {
+        },
+      })((manifestObj) => {
         it(`checks return url`, () => {
           mockConstants({ executionEnvironment: ExecutionEnvironment.StoreClient }, manifestObj);
 
@@ -161,19 +149,16 @@ if (Platform.OS !== 'web') {
         });
       });
 
-      describeManifestTypes(
-        { scheme: 'my-app', hostUri: 'exp.host/@example/abc' },
-        {
-          extra: {
-            expoClient: {
-              name: 'wat',
-              slug: 'wat',
-              scheme: 'my-app',
-              hostUri: 'exp.host/@example/abc',
-            },
+      describeManifest({
+        extra: {
+          expoClient: {
+            name: 'wat',
+            slug: 'wat',
+            scheme: 'my-app',
+            hostUri: 'exp.host/@example/abc',
           },
-        }
-      )((manifestObj) => {
+        },
+      })((manifestObj) => {
         it(`checks return url with options`, () => {
           mockConstants({ executionEnvironment: ExecutionEnvironment.StoreClient }, manifestObj);
 
@@ -188,19 +173,16 @@ if (Platform.OS !== 'web') {
         });
       });
 
-      describeManifestTypes(
-        { scheme: ['my-app-1', 'my-app-2'], hostUri: 'exp.host/@test/test' },
-        {
-          extra: {
-            expoClient: {
-              name: 'exp.host',
-              slug: 'exp.host',
-              scheme: ['my-app-1', 'my-app-2'],
-              hostUri: 'exp.host/@test/test',
-            },
+      describeManifest({
+        extra: {
+          expoClient: {
+            name: 'exp.host',
+            slug: 'exp.host',
+            scheme: ['my-app-1', 'my-app-2'],
+            hostUri: 'exp.host/@test/test',
           },
-        }
-      )((manifestObj) => {
+        },
+      })((manifestObj) => {
         it(`checks return url with multiple schemes and no default provided`, () => {
           mockConstants({ executionEnvironment: ExecutionEnvironment.StoreClient }, manifestObj);
 
@@ -214,19 +196,16 @@ if (Platform.OS !== 'web') {
         });
       });
 
-      describeManifestTypes(
-        { scheme: 'my-app', hostUri: 'exp.host/@example/abc?release-channel=release-channel' },
-        {
-          extra: {
-            expoClient: {
-              name: 'wat',
-              slug: 'wat',
-              scheme: 'my-app',
-              hostUri: 'exp.host/@example/abc?release-channel=release-channel',
-            },
+      describeManifest({
+        extra: {
+          expoClient: {
+            name: 'wat',
+            slug: 'wat',
+            scheme: 'my-app',
+            hostUri: 'exp.host/@example/abc?release-channel=release-channel',
           },
-        }
-      )((manifestObj) => {
+        },
+      })((manifestObj) => {
         it(`checks url with the release channel`, () => {
           mockConstants({ executionEnvironment: ExecutionEnvironment.StoreClient }, manifestObj);
 
@@ -250,19 +229,16 @@ if (Platform.OS !== 'web') {
         });
         afterEach(() => (console.warn = originalWarn));
 
-        describeManifestTypes(
-          { scheme: 'my-app', hostUri: 'exp.host/@example/abc' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                scheme: 'my-app',
-                hostUri: 'exp.host/@example/abc',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              scheme: 'my-app',
+              hostUri: 'exp.host/@example/abc',
             },
-          }
-        )((manifestObj) => {
+          },
+        })((manifestObj) => {
           it(`checks return url`, () => {
             mockConstants({ executionEnvironment: execution }, manifestObj);
 
@@ -275,19 +251,16 @@ if (Platform.OS !== 'web') {
           });
         });
 
-        describeManifestTypes(
-          { scheme: undefined, hostUri: 'exp.host/@test/test' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                scheme: undefined,
-                hostUri: 'exp.host/@test/test',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              scheme: undefined,
+              hostUri: 'exp.host/@test/test',
             },
-          }
-        )((manifestObj) => {
+          },
+        })((manifestObj) => {
           it(`throws if no scheme is defined`, () => {
             mockConstants({ executionEnvironment: execution }, manifestObj);
 
@@ -300,19 +273,16 @@ if (Platform.OS !== 'web') {
           });
         });
 
-        describeManifestTypes(
-          { scheme: 'my-app', hostUri: 'exp.host/@example/abc' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                scheme: 'my-app',
-                hostUri: 'exp.host/@example/abc',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              scheme: 'my-app',
+              hostUri: 'exp.host/@example/abc',
             },
-          }
-        )((manifestObj) => {
+          },
+        })((manifestObj) => {
           it(`checks return url with options`, () => {
             mockConstants({ executionEnvironment: execution }, manifestObj);
 
@@ -331,19 +301,16 @@ if (Platform.OS !== 'web') {
           });
         });
 
-        describeManifestTypes(
-          { scheme: ['my-app-1', 'my-app-2'], hostUri: 'exp.host/@test/test' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                scheme: ['my-app-1', 'my-app-2'],
-                hostUri: 'exp.host/@test/test',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              scheme: ['my-app-1', 'my-app-2'],
+              hostUri: 'exp.host/@test/test',
             },
-          }
-        )((manifestObj) => {
+          },
+        })((manifestObj) => {
           it(`checks return url with multiple schemes and no default provided`, () => {
             mockConstants({ executionEnvironment: execution }, manifestObj);
 
@@ -357,19 +324,16 @@ if (Platform.OS !== 'web') {
           });
         });
 
-        describeManifestTypes(
-          { scheme: 'my-app', hostUri: 'exp.host/@example/abc?release-channel=release-channel' },
-          {
-            extra: {
-              expoClient: {
-                name: 'wat',
-                slug: 'wat',
-                scheme: 'my-app',
-                hostUri: 'exp.host/@example/abc?release-channel=release-channel',
-              },
+        describeManifest({
+          extra: {
+            expoClient: {
+              name: 'wat',
+              slug: 'wat',
+              scheme: 'my-app',
+              hostUri: 'exp.host/@example/abc?release-channel=release-channel',
             },
-          }
-        )((manifestObj) => {
+          },
+        })((manifestObj) => {
           it(`checks url with the release channel`, () => {
             mockConstants({ executionEnvironment: execution }, manifestObj);